services/user-management/pyproject.toml
[tool.poetry]
name = "user-management"
version = "0.1.0"
description = "User Management Microservice for eHealth Platform"
authors = ["Your Name <you@example.com>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
pydantic = {extras = ["email"], version = "^2.7.4"}
pydantic-settings = "^2.3.4"
sqlalchemy = "^2.0.30"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
tenacity = "^8.4.1" # For retrying DB connection
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
redis = "^5.0.5"
boto3 = "^1.34.131"
celery = {extras = ["redis"], version = "^5.4.0"}
python-multipart = "^0.0.9"
pyotp = "^2.9.0"
qrcode = {extras = ["pil"], version = "^7.4.2"}
stripe = "^9.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
httpx = "^0.27.0" # For testing async API
black = "^24.4.2"
isort = "^5.13.2"
mypy = "^1.10.0"
sqlalchemy-utils = "^0.41.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"